Renamed /mint folder to /option-lab
