+++ src/app/option-lab/page.tsx
@@ -1,4 +1,4 @@
-import { OptionLabForm } from "@/components/mint/OptionLabForm"
+import { OptionLabForm } from "@/components/option-lab/OptionLabForm"
 
 export default function MintOptionPage() {
   return (
@@ -6,7 +6,7 @@ export default function MintOptionPage() {
       <div className="mb-8 text-center">
         <h1 className="text-4xl font-thin mb-2">Option Lab</h1>
         <p className="text-muted-foreground">
-          Create, visualize, and mint option contracts to list on the exchange.
+          Create, visualize, and mint option contracts to list on Solana OPX.
         </p>
       </div>
       <OptionLabForm />
