+++ src/components/option-lab/PremiumDisplay.tsx
