+++ src/components/option-lab/QuantityInput.tsx
