+++ src/components/option-lab/StrikePriceInput.tsx
